# specs/007-licensing-service-standalone/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Licensing Service API
  description: |
    Standalone licensing service for issuing, managing, and validating license tokens
    for multiple products. All endpoints require OIDC authentication.
  version: 1.0.0
  contact:
    name: MrWhoOidc Team
servers:
  - url: https://licensing.example.com/api/v1
    description: Production
  - url: https://localhost:5001/api/v1
    description: Local Development
security:
  - bearerAuth: []
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      description: Returns paginated list of customers, optionally filtered by status or search term
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or identifier (prefix match)
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Inactive]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated customer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Customer identifier already exists
  /customers/{customerId}:
    parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Customers]
      summary: Get customer details
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Customers]
      summary: Update customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Customers]
      summary: Soft-delete customer
      description: Sets customer status to Inactive. Fails if customer has active licenses.
      responses:
        '204':
          description: Customer deactivated
        '400':
          description: Customer has active licenses
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Authentication required
    NotFound:
      description: Resource not found
    BadRequest:
      description: Invalid request
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        identifier:
          type: string
        displayName:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
        createdAt:
          type: string
          format: date-time
        licenseCount:
          type: integer
          description: Number of licenses held
    CreateCustomerRequest:
      type: object
      required: [identifier, displayName]
      properties:
        identifier:
          type: string
          pattern: '^[A-Za-z0-9-]+$'
          maxLength: 100
        displayName:
          type: string
          maxLength: 200
        contactEmail:
          type: string
          format: email
        contactName:
          type: string
    UpdateCustomerRequest:
      type: object
      properties:
        displayName:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        status:
          type: string
          enum: [Active, Inactive]
    CustomerListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # ---- Product Schemas ----
...